requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
openai-agents==0.2.4
pydantic==2.11.7
chromadb==1.0.15
markitdown[all]==0.1.2
python-multipart==0.0.10

# Testing dependencies
pytest==8.4.1
pytest-asyncio==0.24.0
httpx==0.28.1
requests==2.32.3
